Update TLineItem scoreMaximum typing.
Align TLineItem.scoreMaximum with float usage in the line item model and AGS spec reference: https://www.imsglobal.org/spec/lti-ags/v2p0#scoremaximum

Made-with: Cursor

+++ pylti1p3/lineitem.py
@@ -17,7 +17,7 @@ class TLineItem(t.TypedDict, total=False):
     """Raw line item payload returned by, or sent to, the platform."""
 
     id: str
-    scoreMaximum: int
+    scoreMaximum: float
     label: str
     resourceId: str
     tag: str
